Changed ProcessingTimeInMS to be double instead of long

+++ src/CAAS/Models/Base/HealthCheckResponse.cs
@@ -13,6 +13,6 @@ namespace CAAS.Models.Base
 
         [Required]
         [DefaultValue(11)]
-        public long ProcessingTimeInMs { get; set; }
+        public double ProcessingTimeInMs { get; set; }
     }
 }
